move functions into helpers file
